:recycle: Refactor: MemeLongPressContainer 교체
