+++ src/components/collect/SearchedCollection/SearchedCollection.tsx
@@ -1,5 +1,5 @@
 import { useGetMemesFromCollectionByKeyword } from "@/application/hooks";
-import { InfiniteMemeList, MemeLongPressContainer } from "@/components/meme";
+import { MasonryInfiniteGrid } from "@/components/meme";
 
 interface Props {
   searchQuery: string;
@@ -19,9 +19,5 @@ export const SearchedCollection = ({ searchQuery, collectionId }: Props) => {
   if (isEmpty) {
     return null;
   }
-  return (
-    <MemeLongPressContainer memeList={memeList}>
-      <InfiniteMemeList memeList={memeList} onEndReached={fetchNextPage} />
-    </MemeLongPressContainer>
-  );
+  return <MasonryInfiniteGrid memeList={memeList} onEndReached={fetchNextPage} />;
 };
