+++ src/components/home/MemeListContainer/MemeList/CommonMemeList.tsx
@@ -1,5 +1,5 @@
 import { useGetMemesBySort } from "@/application/hooks";
-import { InfiniteMemeList, MemeLongPressContainer } from "@/components/meme";
+import { MasonryInfiniteGrid } from "@/components/meme";
 
 import type { MemeListType } from "../type";
 
@@ -9,9 +9,5 @@ interface Props {
 export const CommonMemeList = ({ sortBy }: Props) => {
   const { data: memeList, fetchNextPage } = useGetMemesBySort(sortBy);
 
-  return (
-    <MemeLongPressContainer memeList={memeList}>
-      <InfiniteMemeList memeList={memeList} onEndReached={fetchNextPage} />
-    </MemeLongPressContainer>
-  );
+  return <MasonryInfiniteGrid memeList={memeList} onEndReached={fetchNextPage} />;
 };
