+++ src/components/meme/index.ts
@@ -1,4 +1,4 @@
-export * from "./InfiniteMemeList";
 export * from "./LongPress";
+export * from "./MansonryInfiniteGrid";
 export * from "./MemeInfo";
 export * from "./MemeItem";
